refactor: move network objects to their own crate
