+++ network/README.md
@@ -1,0 +1,2 @@
+# snarkOS-network
+
